Refactor GlyphWidthsMap (#24)
* Replaced internal HashMap<i64, Vec<f32>> with
  BTreeMap<i64, WidthRun> for deterministic
  ordering and efficient predecessor lookups.
* Added WidthRun enum (Explicit, Uniform) to
  avoid materializing large uniform ranges.
* Removed heuristic missing-width logic;
  now strictly follows spec expectations.
* Enforced validation: non-empty explicit arrays,
  no duplicate starts, no overlaps, range size upper bound.
* Rewrote parsing with safe arithmetic (saturating_add,
  checked conversions) and overlap detection.
* Added new tests covering empty array, duplicate starts,
  overlapping definitions (both directions), oversized range,
  and uniform range lookup.

+++ crates/pdf-canvas/src/truetype_font_renderer.rs
@@ -163,7 +163,7 @@ impl<T: PdfOperatorBackend + Canvas> TextRenderer for TrueTypeFontRenderer<'_, T
             let w0_glyph_units = cid_font
                 .widths
                 .as_ref()
-                .and_then(|w_array| w_array.get_width(i64::from(char_code)))
+                .and_then(|w_array| w_array.get_width(char_code))
                 .unwrap_or(cid_font.default_width);
 
             // Convert width from font units to ems.
